+++ dsa/formal_languages/__init__.py
@@ -1,5 +1,5 @@
 from .cnf_tools import chomsky_normal_form, grammar_is_cnf
-from .context_free import CFG
+from .grammar import CFG
 from .cyk import CYKParser
 from .types import Nonterminal, productiontype
